+++ src/fastjsDom/fastjsDom.js
@@ -1,29 +1,26 @@
-import fastjsDom from "../fastjsDom/main";
 import _dev from "../dev";
 import dataEdit from "./dataEdit";
 import domEdit from "./domEdit";
 import event from "./event";
 
-const main = function (list) {
-    let domList = [];
-    list.forEach((v) => {
-        domList.push(new fastjsDom(v));
-    })
+const fastjsDom = function (el) {
+    // if string
+    if (typeof el === "string") {
+        // create
+        el = _dev._dom.createElement(el);
+    }
 
     let _this = {
-        _list: domList,
-        construct: "fastjsDomList"
+        _el: el,
+        construct: "fastjsDom"
     }
 
     // import methods
     _this = _dev.initMethod(_this, dataEdit, domEdit, event);
 
-    domList.forEach((e, key) => {
-        this[key] = e;
-    })
     Object.entries(_this).forEach(e => {
         this[e[0]] = e[1];
     })
 }
 
-export default main
+export default fastjsDom
